starting to break apart the line parser. Broke out two kinds of lines into simple static classes.
